+++ year_one/libft/ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: mmacdona <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/24 13:15:49 by mmacdona          #+#    #+#             */
-/*   Updated: 2017/06/03 14:05:59 by mmacdona         ###   ########.fr       */
+/*   Updated: 2017/06/06 13:27:37 by mmacdona         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strsub(char const *s, unsigned int start, size_t len)
 {
-	int		i;
+	size_t	i;
 	char	*str;
 
 	str = NULL;
